Fix the way to make the installer appear in the output directory.
Before the MoveInstaller target was used to copy file from fake
OutDir to 'bin', but it has some problems; The target sometimes
doesn't want to work and this is not the best solution to solve
the problem. This change uses WIX OutDir parameter to set the
destination path. The destination path of the installer is now
'bin\Release\Installer\en-US\PoshSvn.msi'.

* .github/workflows/release.yml
  * Update path in upload installer to release step from
    'bin\Release\PoshSvn.msi' to 'bin\Release\Installer\en-US\PoshSvn.msi'.
* Installer/Installer.wixproj
  * Change OutDir to bin\Release\Installer
  * Remove MoveInstaller target and its parameters.

git-svn-id: file:///c:/Repositories/rinrab/poshsvn/trunk@1152 0083e6b4-6090-234a-836e-88fcc467669c